refactor: moved Conan to new ToolVersionLogger
