+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/bitbake/BitbakeSession.java
@@ -12,7 +12,7 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
-import com.synopsys.integration.detectable.detectable.util.ToolVersionLogger;
+import com.synopsys.integration.detectable.util.ToolVersionLogger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
