+++ detectable/src/main/java/com/synopsys/integration/detectable/detectables/go/gomod/GoModCommandExecutor.java
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.synopsys.integration.detectable.detectable.util.ToolVersionLogger;
+import com.synopsys.integration.detectable.util.ToolVersionLogger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
